@@ -12,8 +12,8 @@
     ],
     "repository": "gitlab:mpapp-public/prosemirror-recreate-steps",
     "scripts": {
-        "test": "mocha test/test-*.js",
-        "coverage": "nyc --reporter=html --check-coverage --statements 36.11 --branches 18.9 --lines 39.13 --functions 33.87 mocha test/test-*.js",
+        "test": "jest",
+        "coverage": "nyc --reporter=html --check-coverage --statements 36.11 --branches 18.9 --lines 39.13 --functions 33.87 jest",
         "build": "npm run build_js && npm run build_demo_js && npm run build_css",
         "build_demo_js": "npm run build_merge_demo_js && npm run build_merge2_demo_js && npm run build_history_demo_js",
         "build_merge_demo_js": "rollup demo/merge/index.js -c -f iife -o demo/merge/index-build.js",
@@ -32,42 +32,37 @@
     ],
     "author": "Johannes Wilm",
     "license": "Apache-2.0",
-    "dependencies": {
-        "@types/prosemirror-model": "^1.5.0",
+    "devDependencies": {
+        "@babel/preset-env": "^7.7.1",
+        "@manuscripts/publish": "^0.2.2",
+        "@types/prosemirror-model": "^1.7.2",
         "@types/prosemirror-schema-basic": "^1.0.1",
         "@types/prosemirror-schema-list": "^1.0.1",
-        "@types/prosemirror-state": "^1.2.0",
-        "@types/prosemirror-transform": "^1.1.0",
-        "diff": "^3.5.0",
-        "prosemirror-changeset": "^1.0.5",
-        "prosemirror-example-setup": "^1.0.1",
-        "prosemirror-model": "^1.6.1",
-        "prosemirror-schema-basic": "^1.0.0",
-        "prosemirror-schema-list": "^1.0.1",
-        "prosemirror-state": "^1.2.2",
-        "prosemirror-transform": "^1.1.3",
-        "prosemirror-view": "^1.5.3",
-        "rfc6902": "^3.0.1"
-    },
-    "devDependencies": {
-        "@manuscripts/publish": "^0.2.1",
-        "eslint": "^5.6.0",
-        "eslint-config-standard": "^12.0.0",
-        "eslint-plugin-import": "^2.14.0",
-        "eslint-plugin-node": "^7.0.1",
-        "eslint-plugin-promise": "^4.0.1",
-        "eslint-plugin-standard": "^4.0.0",
+        "@types/prosemirror-state": "^1.2.3",
+        "@types/prosemirror-transform": "^1.1.1",
+        "diff": "^4.0.1",
+        "eslint": "^6.6.0",
+        "eslint-config-standard": "^14.1.0",
+        "eslint-plugin-import": "^2.18.2",
+        "eslint-plugin-node": "^10.0.0",
+        "eslint-plugin-promise": "^4.2.1",
+        "eslint-plugin-standard": "^4.0.1",
         "ist": "^1.1.1",
-        "jshint": "^2.9.6",
-        "mocha": "^5.2.0",
-        "nyc": "^13.0.1",
+        "jest": "^24.9.0",
+        "jshint": "^2.10.3",
+        "nyc": "^14.1.1",
+        "prosemirror-changeset": "^2.1.0",
+        "prosemirror-example-setup": "^1.1.0",
+        "prosemirror-model": "^1.8.0",
+        "prosemirror-schema-basic": "^1.1.0",
+        "prosemirror-schema-list": "^1.1.0",
+        "prosemirror-state": "^1.3.0",
         "prosemirror-test-builder": "^1.0.1",
-        "rollup": "^0.66.2",
-        "rollup-plugin-commonjs": "^9.1.8",
-        "rollup-plugin-node-resolve": "^3.4.0"
-    },
-    "jshintConfig": {
-        "esversion": 6,
-        "asi": true
+        "prosemirror-transform": "^1.2.0",
+        "prosemirror-view": "^1.13.2",
+        "rfc6902": "^3.0.4",
+        "rollup": "^1.27.0",
+        "rollup-plugin-commonjs": "^10.1.0",
+        "rollup-plugin-node-resolve": "^5.2.0"
     }
 }
