@@ -11,7 +11,7 @@ import {
 export function mergeTransforms(tr1, tr2, automerge = true, rebase = false, wordDiffs = false) {
     // Create conflicting steps. Make sure the steps are only ReplaceSteps so they can easily
     // be presented as alternatives to the user.
-    const {tr, changes, tr1NoConflicts, tr2NoConflicts} =
+    const { tr, changes, tr1NoConflicts, tr2NoConflicts } =
         automerge
             ? automergeTransforms(tr1, tr2)
             : noAutomergeTransforms(tr1, tr2),
@@ -26,8 +26,8 @@ export function mergeTransforms(tr1, tr2, automerge = true, rebase = false, word
             ),
             tr.doc,
             new Mapping(tr1NoConflicts.mapping.invert().maps.concat(tr.mapping.maps))
-        )
-    const tr2Conflict = mapTransform(
+        ),
+        tr2Conflict = mapTransform(
             recreateTransform(
                 tr2NoConflicts.doc,
                 tr2.doc,
@@ -43,19 +43,18 @@ export function mergeTransforms(tr1, tr2, automerge = true, rebase = false, word
         return rebaseMergedTransform(tr1.doc, tr1Conflict.doc, tr2Conflict.doc, wordDiffs)
     } else {
         const conflicts = findConflicts(tr1Conflict, tr2Conflict),
-            {inserted, deleted, conflictingSteps1, conflictingSteps2} = createConflictingChanges(tr1Conflict, tr2Conflict)
+            { conflictingChanges, conflictingSteps1, conflictingSteps2 } = createConflictingChanges(tr1Conflict, tr2Conflict)
 
-        return {tr, merge: new Merge(tr.doc, changes, conflicts, conflictingSteps1, conflictingSteps2, {inserted, deleted})}
+        return { tr, merge: new Merge(tr.doc, changes, conflicts, conflictingSteps1, conflictingSteps2, conflictingChanges) }
     }
 }
 
 function rebaseMergedTransform(doc, nonConflictingDoc, conflictingDoc, wordDiffs) {
     const trNonConflict = recreateTransform(doc, nonConflictingDoc, true, wordDiffs),
-        changes = ChangeSet.create(doc, {compare: (a, b) => false}).addSteps(nonConflictingDoc, trNonConflict.mapping.maps, {user: 2}),
+        changes = ChangeSet.create(doc).addSteps(nonConflictingDoc, trNonConflict.mapping.maps, { user: 2 }),
         trConflict = recreateTransform(nonConflictingDoc, conflictingDoc, false, wordDiffs),
         {
-            inserted,
-            deleted,
+            conflictingChanges,
             conflictingSteps2
         } = createConflictingChanges(
             new Transform(trNonConflict.doc),
@@ -65,7 +64,7 @@ function rebaseMergedTransform(doc, nonConflictingDoc, conflictingDoc, wordDiffs
     return {
         tr: trNonConflict,
         merge: new Merge(
-            trNonConflict.doc, changes, [], [], conflictingSteps2, {inserted, deleted}
+            trNonConflict.doc, changes, [], [], conflictingSteps2, conflictingChanges
         )
     }
 }
@@ -77,7 +76,7 @@ export class Merge {
         conflicts = [],
         conflictingSteps1 = [],
         conflictingSteps2 = [],
-        conflictingChanges = {inserted: [], deleted: []}
+        conflictingChanges = []
     ) {
         this.doc = doc
         this.changes = changes
@@ -91,25 +90,24 @@ export class Merge {
         let conflictingSteps1 = this.conflictingSteps1,
             conflictingSteps2 = this.conflictingSteps2,
             conflicts = this.conflicts,
-            inserted = this.conflictingChanges.inserted,
-            deleted = this.conflictingChanges.deleted
-        const changes = this.changes.addSteps(doc, mapping.maps, {user: 2})
+            conflictingChanges = this.conflictingChanges
+        const changes = this.changes.addSteps(doc, mapping.maps, { user: 2 })
 
         conflictingSteps1 = conflictingSteps1.map(
             ([index, conflictStep]) => {
                 const mapped = conflictStep.map(mapping)
                 if (mapped) {
-                    inserted = inserted.map(
-                        inserted => ({data: inserted.data, slice: inserted.slice, pos: mapping.map(inserted.pos)})
-                    )
-                    deleted = deleted.map(
-                        deleted => ({data: deleted.data, from: mapping.map(deleted.from), to: mapping.map(deleted.to)})
+                    conflictingChanges = conflictingChanges.map(
+                        change => ({ inserted: change.inserted, deleted: change.deleted, fromA: mapping.map(change.fromA), toA: mapping.map(change.toA), fromB: mapping.map(change.fromB), toB: mapping.map(change.toB) })
                     )
                     return [index, mapped]
                 } else {
                     conflicts = conflicts.filter(conflict => conflict[0] !== index)
-                    inserted = inserted.filter(inserted => inserted.data.user !== 1 || inserted.data.index !== index)
-                    deleted = deleted.filter(deleted => deleted.data.user !== 1 || deleted.data.index !== index)
+                    conflictingChanges = conflictingChanges.filter(change => {
+                        change.deleted = change.deleted.filter(deleted => deleted.data.user !== 1 || deleted.data.index !== index)
+                        change.inserted = change.inserted.filter(inserted => inserted.data.user !== 1 || inserted.data.index !== index)
+                        return change.deleted.length || change.inserted.length
+                    })
                     return false
                 }
             }
@@ -119,23 +117,23 @@ export class Merge {
             ([index, conflictStep]) => {
                 const mapped = conflictStep.map(mapping)
                 if (mapped) {
-                    inserted = inserted.map(
-                        inserted => ({data: inserted.data, slice: inserted.slice, pos: mapping.map(inserted.pos)})
-                    )
-                    deleted = deleted.map(
-                        deleted => ({data: deleted.data, from: mapping.map(deleted.from), to: mapping.map(deleted.to)})
+                    conflictingChanges = conflictingChanges.map(
+                        change => ({ inserted: change.inserted, deleted: change.deleted, fromA: mapping.map(change.fromA), toA: mapping.map(change.toA), fromB: mapping.map(change.fromB), toB: mapping.map(change.toB) })
                     )
                     return [index, mapped]
                 } else {
                     conflicts = conflicts.filter(conflict => conflict[1] !== index)
-                    inserted = inserted.filter(inserted => inserted.data.user !== 2 || inserted.data.index !== index)
-                    deleted = deleted.filter(deleted => deleted.data.user !== 2 || deleted.data.index !== index)
+                    conflictingChanges = conflictingChanges.filter(change => {
+                        change.deleted = change.deleted.filter(deleted => deleted.data.user !== 2 || deleted.data.index !== index)
+                        change.inserted = change.inserted.filter(inserted => inserted.data.user !== 2 || inserted.data.index !== index)
+                        return change.deleted.length || change.inserted.length
+                    })
                     return false
                 }
             }
         ).filter(step => step)
 
-        return new Merge(doc, changes, conflicts, conflictingSteps1, conflictingSteps2, {inserted, deleted})
+        return new Merge(doc, changes, conflicts, conflictingSteps1, conflictingSteps2, conflictingChanges)
     }
 
     apply(user, index) {
@@ -150,7 +148,7 @@ export class Merge {
 
         tr.step(step)
 
-        const changes = this.changes.addSteps(tr.doc, [map], {user})
+        const changes = this.changes.addSteps(tr.doc, [map], { user })
 
         if (user === 1) {
             conflictingSteps1 = conflictingSteps1.map(
@@ -164,16 +162,13 @@ export class Merge {
             conflicts = conflicts.filter(conflict => conflict[1] !== index)
         }
 
-        const conflictingChanges = {
-            inserted: this.conflictingChanges.inserted.filter(inserted => inserted.data.user !== user || inserted.data.index !== index).map(
-                inserted => ({data: inserted.data, slice: inserted.slice, pos: map.map(inserted.pos)})
-            ),
-            deleted: this.conflictingChanges.deleted.filter(deleted => deleted.data.user !== user || deleted.data.index !== index).map(
-                deleted => ({data: deleted.data, from: map.map(deleted.from), to: map.map(deleted.to)})
-            )
-        }
+        const conflictingChanges = conflictingChanges.filter(change => {
+            change.deleted = change.deleted.filter(deleted => deleted.data.user !== user || deleted.data.index !== index)
+            change.inserted = change.inserted.filter(inserted => inserted.data.user !== user || inserted.data.index !== index)
+            return change.deleted.length || change.inserted.length
+        })
 
-        return {tr, merge: new Merge(tr.doc, changes, conflicts, conflictingSteps1, conflictingSteps2, conflictingChanges)}
+        return { tr, merge: new Merge(tr.doc, changes, conflicts, conflictingSteps1, conflictingSteps2, conflictingChanges) }
     }
 
     reject(user, index) {
@@ -193,12 +188,13 @@ export class Merge {
             conflicts = conflicts.filter(conflict => conflict[1] !== index)
         }
 
-        const conflictingChanges = {
-            inserted: this.conflictingChanges.inserted.filter(inserted => inserted.data.user !== user || inserted.data.index !== index),
-            deleted: this.conflictingChanges.deleted.filter(deleted => deleted.data.user !== user || deleted.data.index !== index)
-        }
+        const conflictingChanges = conflictingChanges.filter(change => {
+            change.deleted = change.deleted.filter(deleted => deleted.data.user !== user || deleted.data.index !== index)
+            change.inserted = change.inserted.filter(inserted => inserted.data.user !== user || inserted.data.index !== index)
+            return change.deleted.length || change.inserted.length
+        })
 
-        return {merge: new Merge(this.doc, this.changes, conflicts, conflictingSteps1, conflictingSteps2, conflictingChanges)}
+        return { merge: new Merge(this.doc, this.changes, conflicts, conflictingSteps1, conflictingSteps2, conflictingChanges) }
     }
 
     applyAll(user) {
@@ -208,10 +204,10 @@ export class Merge {
         while (steps.length) {
             const mapped = steps.pop().map(tr.mapping)
             if (mapped && !tr.maybeStep(mapped).failed) {
-                changes = this.changes.addSteps(tr.doc, [tr.mapping.maps[tr.mapping.maps.length - 1]], {user})
+                changes = this.changes.addSteps(tr.doc, [tr.mapping.maps[tr.mapping.maps.length - 1]], { user })
             }
         }
-        return {tr, merge: new Merge(tr.doc, changes)}
+        return { tr, merge: new Merge(tr.doc, changes) }
     }
 }
 
@@ -224,7 +220,7 @@ function mapTransform(tr, doc, map) {
             try {
                 newTr.maybeStep(mapped)
             } catch (error) {
-                if(!error.name === 'ReplaceError') {
+                if (!error.name === 'ReplaceError') {
                     throw error
                 }
             }
@@ -241,7 +237,7 @@ function noAutomergeTransforms(tr1, tr2) {
     const doc = trDoc(tr1)
     return {
         tr: new Transform(doc),
-        changes: ChangeSet.create(doc, {compare: (a, b) => false}),
+        changes: ChangeSet.create(doc, { compare: (a, b) => false }),
         tr1NoConflicts: new Transform(doc),
         tr2NoConflicts: new Transform(doc)
     }
@@ -252,22 +248,22 @@ function automergeTransforms(tr1, tr2) {
     const doc = trDoc(tr1),
         conflicts = findConflicts(tr1, tr2),
         tr = new Transform(doc)
-    let changes = ChangeSet.create(doc, {compare: (a, b) => false})
+    let changes = ChangeSet.create(doc, { compare: (a, b) => false })
     const tr1NoConflicts = removeConflictingSteps(tr1, conflicts.map(conflict => conflict[0])),
         tr2NoConflicts = removeConflictingSteps(tr2, conflicts.map(conflict => conflict[1]))
 
     tr1NoConflicts.steps.forEach(step => tr.maybeStep(step))
     const numberSteps1 = tr.steps.length
-    changes = changes.addSteps(tr.doc, tr.mapping.maps, {user: 1})
+    changes = changes.addSteps(tr.doc, tr.mapping.maps, { user: 1 })
     tr2NoConflicts.steps.forEach(step => {
         const mapped = step.map(tr.mapping.slice(0, numberSteps1))
         if (mapped) {
             tr.maybeStep(mapped)
         }
     })
-    changes = changes.addSteps(tr.doc, tr.mapping.maps.slice(numberSteps1), {user: 2})
+    changes = changes.addSteps(tr.doc, tr.mapping.maps.slice(numberSteps1), { user: 2 })
 
-    return {tr, changes, tr1NoConflicts, tr2NoConflicts}
+    return { tr, changes, tr1NoConflicts, tr2NoConflicts }
 }
 
 function removeConflictingSteps(tr, conflicts) {
@@ -292,39 +288,17 @@ function findConflicts(tr1, tr2) {
     const changes1 = findContentChanges(tr1),
         changes2 = findContentChanges(tr2),
         conflicts = []
-    changes1.deleted.forEach(deleted => {
-        changes2.inserted.forEach(inserted => {
-            if (inserted.pos >= deleted.from && inserted.pos <= deleted.to) {
-                conflicts.push([deleted.data.step, inserted.data.step])
-            }
-        })
-    })
-
-    changes2.deleted.forEach(deleted => {
-        changes1.inserted.forEach(inserted => {
-            if (inserted.pos >= deleted.from && inserted.pos <= deleted.to) {
-                conflicts.push([inserted.data.step, deleted.data.step])
-            }
-        })
-    })
-
-    changes1.inserted.forEach(inserted1 => {
-        changes2.inserted.forEach(inserted2 => {
-            if (inserted1.pos === inserted2.pos) {
-                conflicts.push([inserted1.data.step, inserted2.data.step])
-            }
-        })
-    })
-
-    changes1.deleted.forEach(deleted1 => {
-        changes2.deleted.forEach(deleted2 => {
-            if (
-                (deleted1.from >= deleted2.from && deleted1.from <= deleted2.to) ||
-                (deleted1.to >= deleted2.from && deleted1.to <= deleted2.to) ||
-                (deleted1.from <= deleted2.from && deleted1.to >= deleted2.to) ||
-                (deleted2.from <= deleted1.from && deleted2.to >= deleted1.to)
-            ) {
-                conflicts.push([deleted1.data.step, deleted2.data.step])
+    changes1.forEach(change1 => {
+        changes2.forEach(change2 => {
+            if (!(change1.toA < change2.fromA || change1.fromA > change2.toA)) {
+                change1.inserted.forEach(span1 => {
+                    change2.inserted.forEach(span2 => conflicts.push([span1.data.step, span2.data.step]))
+                    change2.deleted.forEach(span2 => conflicts.push([span1.data.step, span2.data.step]))
+                })
+                change1.deleted.forEach(span1 => {
+                    change2.inserted.forEach(span2 => conflicts.push([span1.data.step, span2.data.step]))
+                    change2.deleted.forEach(span2 => conflicts.push([span1.data.step, span2.data.step]))
+                })
             }
         })
     })
@@ -334,17 +308,13 @@ function findConflicts(tr1, tr2) {
 
 function findContentChanges(tr) {
     const doc = trDoc(tr)
-    let changes = ChangeSet.create(doc, {compare: (a, b) => false})
+    let changes = ChangeSet.create(doc)
     tr.steps.forEach((step, index) => {
         const doc = trDoc(tr, index + 1)
-        changes = changes.addSteps(doc, [tr.mapping.maps[index]], {step: index})
+        changes = changes.addSteps(doc, [tr.mapping.maps[index]], { step: index })
     })
-    const invertedMapping = new Mapping()
-    invertedMapping.appendMappingInverted(tr.mapping)
-    const inserted = changes.inserted.map(inserted => ({pos: invertedMapping.map(inserted.from), data: inserted.data}))
-    const deleted = changes.deleted.map(deleted => ({from: deleted.from, to: deleted.to, data: deleted.data}))
 
-    return {inserted, deleted}
+    return changes.changes
 }
 
 function createConflictingChanges(tr1Conflict, tr2Conflict) {
@@ -354,14 +324,13 @@ function createConflictingChanges(tr1Conflict, tr2Conflict) {
         // If one of them is being applied, the other ones will have to be remapped.
         conflictingSteps1 = tr1Conflict.steps.map((step, index) => [index, step.map(new Mapping(tr1Conflict.mapping.maps.slice(0, index)).invert())]),
         conflictingSteps2 = tr2Conflict.steps.map((step, index) => [index, step.map(new Mapping(tr2Conflict.mapping.maps.slice(0, index)).invert())])
-    let inserted = [],
-        deleted = []
+    let conflictingChanges = []
     const iter = [
-        {steps: conflictingSteps1, user: 1},
-        {steps: conflictingSteps2, user: 2}
+        { steps: conflictingSteps1, user: 1 },
+        { steps: conflictingSteps2, user: 2 }
     ]
 
-    iter.forEach(({steps, user}) =>
+    iter.forEach(({ steps, user }) =>
         steps.forEach(([index, step]) => {
             if (!step) {
                 return
@@ -369,10 +338,9 @@ function createConflictingChanges(tr1Conflict, tr2Conflict) {
             const stepResult = step.apply(doc)
             // We need the potential changes if this step was to be applied. We find
             // the inversion of the change so that we can place it in the current doc.
-            const invertedStepChanges = ChangeSet.create(stepResult.doc, {compare: (a, b) => false}).addSteps(doc, [step.invert(doc).getMap()], {index, user})
-            deleted = deleted.concat(invertedStepChanges.inserted.map(inserted => ({from: inserted.from, to: inserted.to, data: inserted.data})))
-            inserted = inserted.concat(invertedStepChanges.deleted.map(deleted => ({pos: deleted.pos, slice: deleted.slice, data: deleted.data})))
+            const invertedStepChanges = ChangeSet.create(stepResult.doc).addSteps(doc, [step.invert(doc).getMap()], { index, user })
+            conflictingChanges = conflictingChanges.concat(invertedStepChanges.changes)
         })
     )
-    return {inserted, deleted, conflictingSteps1, conflictingSteps2}
+    return { conflictingChanges, conflictingSteps1, conflictingSteps2 }
 }
